+++ R/exec_version.R
@@ -10,6 +10,7 @@
 #' * `node`: 'node.js' executable
 #' * `npm`: 'npm' executable
 #' * `air`: 'air' R formatter and language server executable
+#' * `pandoc`: 'Pandoc' document converter executable
 #' * `python`: 'python' executable
 #' * `pip`: 'pip' executable
 #' * `excel`: 'Microsoft Excel' executable
@@ -24,6 +25,8 @@
 #'   message(exec_version("node"))
 #' if(exec_available("air"))
 #'   message(exec_version("air"))
+#' if(exec_available("pandoc"))
+#'   message(exec_version("pandoc"))
 #' if(exec_available("python"))
 #'   message(exec_version("python"))
 #' if(exec_available("libreoffice"))