+++ R/utils.R
@@ -11,6 +11,7 @@
 .exec$powerpoint <- list(exec_file = NULL, version = NULL)
 .exec$word <- list(exec_file = NULL, version = NULL)
 .exec$air <- list(exec_file = NULL, version = NULL)
+.exec$pandoc <- list(exec_file = NULL, version = NULL)
 
 .exec_names <- new.env()
 .exec_names$node <- list(exec_win = "node.exe", exec_osx = "node", exec_unix = "node")
@@ -26,6 +27,7 @@
 .exec_names$powerpoint <- list(exec_win = "POWERPNT.exe", exec_osx = "Microsoft PowerPoint", exec_unix = NA_character_)
 .exec_names$word <- list(exec_win = "WINWORD.exe", exec_osx = "Microsoft Word", exec_unix = NA_character_)
 .exec_names$air <- list(exec_win = "air.exe", exec_osx = "air", exec_unix = "air")
+.exec_names$pandoc <- list(exec_win = "pandoc.exe", exec_osx = "pandoc", exec_unix = "pandoc")
 
 supported_exec <- names(.exec)